ProfileService/requirements.txt
fastapi
uvicorn[standard]
pyodbc
requests
python-dotenv
email-validator
PyJWT